backend/pyproject.toml
[tool.poetry]
name = "edison-ai"
version = "0.1.0"
description = ""
authors = ["shrey shah <shrey.shah@vivun.com>"]
readme = "README.md"

[project]
name = "edison-ai"
version = "0.0.1"
dependencies = [
  "langgraph",
  "langchain_core",
  "langchain_openai",
  "langchain",
  "openai",
  "langchain-community",
  "copilotkit",
  "uvicorn",
  "python-dotenv",
  "bs4",
  "rich",
  "aiosqlite",
  "langchain-anthropic",
]

[tool.poetry.dependencies]
python = "3.12.2"
langgraph = "^0.2.34"
langchain-core = "^0.3.8"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "^0.1.22"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
rich = "^13.9.4"
selenium = "^4.26.1"
aiosqlite = "^0.20.0"
youtube-transcript-api = "^0.6.2"




[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.scripts]
edison-ai = "edison_ai.app:main"
